Review process definition and instance details views
Remove Errai warnings regarding wrong element between UI and view binding
Remove duplicated SLA text from instance details
Refactor view classes to hot expose widget elements
Replace widgets with Elemental2 elements
Remove unnecessary abstract classes
