+++ jbpm-wb-process-runtime/jbpm-wb-process-runtime-client/src/main/java/org/jbpm/workbench/pr/client/editors/definition/details/ProcessDefinitionDetailsTabViewImpl.html
@@ -1,0 +1,69 @@
+<div class="row" style="margin-top: 20px">
+    <div class="col-md-12">
+        <div class="form-horizontal">
+            <div class="form-group">
+                <label class="control-label col-md-2" for="processIdText" data-field="processIdLabel"></label>
+
+                <div class="col-md-10">
+                    <p data-field="processIdText" id="processIdText" class="form-control-static"></p>
+                </div>
+            </div>
+            <div class="form-group">
+                <label class="control-label col-md-2" for="processNameText" data-field="processNameLabel"></label>
+
+                <div class="col-md-10">
+                    <p data-field="processNameText" id="processNameText" class="form-control-static"></p>
+                </div>
+            </div>
+            <div class="form-group">
+                <label class="control-label col-md-2" for="deploymentIdText" data-field="deploymentIdLabel"></label>
+
+                <div class="col-md-10">
+                    <p data-field="deploymentIdText" id="deploymentIdText" class="form-control-static"></p>
+                </div>
+            </div>
+            <div class="form-group">
+                <label class="control-label col-md-2" for="humanTasksListBox" data-field="humanTasksListLabel"></label>
+
+                <div class="col-md-10">
+                    <p data-field="humanTasksListBox" id="humanTasksListBox" class="form-control-static"></p>
+                </div>
+            </div>
+            <div class="form-group">
+                <label class="control-label col-md-2" for="numberOfHumanTasksText" data-field="nroOfHumanTasksLabel"></label>
+
+                <div class="col-md-10">
+                    <p data-field="numberOfHumanTasksText" id="numberOfHumanTasksText" class="form-control-static"></p>
+                </div>
+            </div>
+            <div class="form-group">
+                <label class="control-label col-md-2" for="usersGroupsListBox" data-field="usersGroupsListLabel"></label>
+
+                <div class="col-md-10">
+                    <p data-field="usersGroupsListBox" id="usersGroupsListBox" class="form-control-static"></p>
+                </div>
+            </div>
+            <div class="form-group">
+                <label class="control-label col-md-2" for="subprocessListBox" data-field="subprocessListLabel"></label>
+
+                <div class="col-md-10">
+                    <p data-field="subprocessListBox" id="subprocessListBox" class="form-control-static"></p>
+                </div>
+            </div>
+            <div class="form-group">
+                <label class="control-label col-md-2" for="processDataListBox" data-field="processDataListLabel"></label>
+
+                <div class="col-md-10">
+                    <p data-field="processDataListBox" id="processDataListBox" class="form-control-static"></p>
+                </div>
+            </div>
+            <div class="form-group">
+                <label class="control-label col-md-2" for="processServicesListBox" data-field="processServicesListLabel"></label>
+
+                <div class="col-md-10">
+                    <p data-field="processServicesListBox" id="processServicesListBox" class="form-control-static"></p>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
