+++ jbpm-wb-process-runtime/jbpm-wb-process-runtime-client/src/main/java/org/jbpm/workbench/pr/client/editors/definition/details/ProcessDefinitionDetailsViewImpl.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2018 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,25 +14,23 @@
  * limitations under the License.
  */
 
-package org.jbpm.workbench.pr.client.editors.definition.details.multi.advance;
+package org.jbpm.workbench.pr.client.editors.definition.details;
 
 import javax.enterprise.context.Dependent;
 
 import com.google.gwt.core.client.GWT;
 import com.google.gwt.uibinder.client.UiBinder;
 import com.google.gwt.uibinder.client.UiField;
-import com.google.gwt.user.client.ui.IsWidget;
+import com.google.gwt.user.client.ui.Composite;
 import com.google.gwt.user.client.ui.Widget;
 import org.gwtbootstrap3.client.ui.NavTabs;
 import org.gwtbootstrap3.client.ui.TabContent;
 import org.gwtbootstrap3.client.ui.TabListItem;
 import org.gwtbootstrap3.client.ui.TabPane;
-import org.jbpm.workbench.pr.client.editors.definition.details.multi.BaseProcessDefDetailsMultiViewImpl;
 import org.jbpm.workbench.pr.client.resources.i18n.Constants;
 
 @Dependent
-public class AdvancedProcessDefDetailsMultiViewImpl extends BaseProcessDefDetailsMultiViewImpl
-        implements AdvancedProcessDefDetailsMultiPresenter.AdvancedProcessDefDetailsMultiView {
+public class ProcessDefinitionDetailsViewImpl extends Composite implements ProcessDefinitionDetailsPresenter.ProcessDefinitionDetailsView {
 
     private static Binder uiBinder = GWT.create(Binder.class);
 
@@ -42,7 +40,7 @@ public class AdvancedProcessDefDetailsMultiViewImpl extends BaseProcessDefDetail
     @UiField
     TabContent tabContent;
 
-    private AdvancedProcessDefDetailsMultiPresenter presenter;
+    private ProcessDefinitionDetailsPresenter presenter;
 
     private TabPane definitionDetailsPane;
 
@@ -53,7 +51,7 @@ public class AdvancedProcessDefDetailsMultiViewImpl extends BaseProcessDefDetail
     private TabPane diagramPane;
 
     @Override
-    public void init(final AdvancedProcessDefDetailsMultiPresenter presenter) {
+    public void init(final ProcessDefinitionDetailsPresenter presenter) {
         this.presenter = presenter;
         initWidget(uiBinder.createAndBindUi(this));
         initTabs();
@@ -61,7 +59,7 @@ public class AdvancedProcessDefDetailsMultiViewImpl extends BaseProcessDefDetail
 
     protected void initTabs() {
         definitionDetailsPane = new TabPane() {{
-            add(getTabView());
+            add(presenter.getDetailsView());
             setActive(true);
         }};
         definitionDetailsTab = new TabListItem(Constants.INSTANCE.Definition_Details()) {{
@@ -86,18 +84,7 @@ public class AdvancedProcessDefDetailsMultiViewImpl extends BaseProcessDefDetail
         }
     }
 
-    @Override
-    protected IsWidget getTabView() {
-        return presenter.getTabView();
-    }
-
-    @Override
-    protected void closeDetails() {
-        presenter.closeDetails();
-    }
-
-    interface Binder extends
-                     UiBinder<Widget, AdvancedProcessDefDetailsMultiViewImpl> {
+    interface Binder extends UiBinder<Widget, ProcessDefinitionDetailsViewImpl> {
 
     }
 }
