--- jbpm-wb-process-runtime/jbpm-wb-process-runtime-client/src/main/java/org/jbpm/workbench/pr/client/editors/definition/details/advance/AdvancedProcessDefDetailsViewImpl.html
@@ -1,69 +1,0 @@
-<div class="row" style="margin-top: 20px">
-    <div class="col-md-12">
-        <div class="form-horizontal">
-            <div class="form-group">
-                <strong><label class="control-label col-md-2" for="processIdText" data-field="processIdLabel"></label></strong>
-
-                <div class="col-md-10">
-                    <label data-field="processIdText" id="processIdText" class="form-control-static"></label>
-                </div>
-            </div>
-            <div class="form-group">
-                <strong><label class="control-label col-md-2" for="processNameText" data-field="processNameLabel"></label></strong>
-
-                <div class="col-md-10">
-                    <label data-field="processNameText" id="processNameText" class="form-control-static"></label>
-                </div>
-            </div>
-            <div class="form-group">
-                <strong><label class="control-label col-md-2" for="deploymentIdText" data-field="deploymentIdLabel"></label></strong>
-
-                <div class="col-md-10">
-                    <label data-field="deploymentIdText" id="deploymentIdText" class="form-control-static"></label>
-                </div>
-            </div>
-            <div class="form-group">
-                <strong><label class="control-label col-md-2" for="humanTasksListBox" data-field="humanTasksListLabel"></label></strong>
-
-                <div class="col-md-10">
-                    <label data-field="humanTasksListBox" id="humanTasksListBox" class="form-control-static"></label>
-                </div>
-            </div>
-            <div class="form-group">
-                <strong><label class="control-label col-md-2" for="nroOfHumanTasksText" data-field="nroOfHumanTasksLabel"></label></strong>
-
-                <div class="col-md-10">
-                    <label data-field="nroOfHumanTasksText" id="nroOfHumanTasksText" class="form-control-static"></label>
-                </div>
-            </div>
-            <div class="form-group">
-                <strong><label class="control-label col-md-2" for="usersGroupsListBox" data-field="usersGroupsListLabel"></label></strong>
-
-                <div class="col-md-10">
-                    <label data-field="usersGroupsListBox" id="usersGroupsListBox" class="form-control-static"></label>
-                </div>
-            </div>
-            <div class="form-group">
-                <strong><label class="control-label col-md-2" for="subprocessListBox" data-field="subprocessListLabel"></label></strong>
-
-                <div class="col-md-10">
-                    <label data-field="subprocessListBox" id="subprocessListBox" class="form-control-static"></label>
-                </div>
-            </div>
-            <div class="form-group">
-                <strong><label class="control-label col-md-2" for="processDataListBox" data-field="processDataListLabel"></label></strong>
-
-                <div class="col-md-10">
-                    <label data-field="processDataListBox" id="processDataListBox" class="form-control-static"></label>
-                </div>
-            </div>
-            <div class="form-group">
-                <strong><label class="control-label col-md-2" for="processServicesListBox" data-field="processServicesListLabel"></label></strong>
-
-                <div class="col-md-10">
-                    <label data-field="processServicesListBox" id="processServicesListBox" class="form-control-static"></label>
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
