+++ jbpm-wb-process-runtime/jbpm-wb-process-runtime-client/src/main/java/org/jbpm/workbench/pr/client/editors/instance/details/ProcessInstanceDetailsTabViewImpl.html
@@ -39,13 +39,6 @@
                     </div>
                 </div>
                 <div class="form-group">
-                    <label class="control-label col-md-2" for="slaComplianceText" data-field="slaComplianceLabel"></label>
-
-                    <div class="col-md-10">
-                        <p class="form-control-static" data-field="slaComplianceText" id="slaComplianceText"/>
-                    </div>
-                </div>
-                <div class="form-group">
                     <label class="control-label col-md-2" for="correlationKeyText" data-field="correlationKeyLabel"></label>
 
                     <div class="col-md-10">
