+++ jbpm-wb-process-runtime/jbpm-wb-process-runtime-client/src/main/java/org/jbpm/workbench/pr/client/editors/instance/details/ProcessInstanceDetailsViewImpl.ui.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!DOCTYPE ui:UiBinder SYSTEM "http://dl.google.com/gwt/DTD/xhtml.ent">
+<ui:UiBinder xmlns:ui="urn:ui:com.google.gwt.uibinder"
+             xmlns:b="urn:import:org.gwtbootstrap3.client.ui">
+
+  <b:Container fluid="true">
+    <b:Row>
+      <b:Column size="MD_12">
+        <b:TabPanel addStyleNames="uf-tabbar-panel">
+          <b:NavTabs ui:field="navTabs" addStyleNames="uf-tabbar-panel-nav-tabs nav-tabs-pf"/>
+          <b:TabContent ui:field="tabContent"/>
+        </b:TabPanel>
+      </b:Column>
+    </b:Row>
+  </b:Container>
+
+</ui:UiBinder>
