--- jbpm-wb-process-runtime/jbpm-wb-process-runtime-client/src/main/java/org/jbpm/workbench/pr/client/editors/instance/details/multi/ProcessInstanceDetailsMultiViewImpl.ui.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-~ Copyright 2014 Red Hat, Inc. and/or its affiliates.
-~
-~ Licensed under the Apache License, Version 2.0 (the "License");
-~ you may not use this file except in compliance with the License.
-~ You may obtain a copy of the License at
-~
-~      http://www.apache.org/licenses/LICENSE-2.0
-~
-~ Unless required by applicable law or agreed to in writing, software
-~ distributed under the License is distributed on an "AS IS" BASIS,
-~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-~ See the License for the specific language governing permissions and
-~ limitations under the License.
--->
-<!DOCTYPE ui:UiBinder SYSTEM "http://dl.google.com/gwt/DTD/xhtml.ent">
-<ui:UiBinder xmlns:ui="urn:ui:com.google.gwt.uibinder"
-             xmlns:b="urn:import:org.gwtbootstrap3.client.ui">
-
-    <b:Container fluid="true">
-        <b:Row>
-            <b:Column size="MD_12">
-                <b:TabPanel addStyleNames="uf-tabbar-panel">
-                    <b:NavTabs ui:field="navTabs" addStyleNames="uf-tabbar-panel-nav-tabs nav-tabs-pf"/>
-                    <b:TabContent ui:field="tabContent"/>
-                </b:TabPanel>
-            </b:Column>
-        </b:Row>
-    </b:Container>
-
-</ui:UiBinder>
